pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mopi"
version = "0.0.1"
authors = [
  { name="Mark Szulyovszky", email="mark.szulyovszky@gmail.com" },
  { name="Daniel Szemerey", email="daniel@szemerey.eu" },
]
description = "Modular Pipelines (mopi) let's you create and manage Machine Learning pipelines with ease."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]

dependencies = [
  "pyarrow>=3.0",
  "black",
  "ipython",
  "notebook",
  "ipykernel",
  "pytorch",
  "tqdm",
  "numpy",
  "pandas",
  "pytest",
  "tensorboard",
  "pyright",
  "wandb",
  "python-dotenv",
  "nltk",
  "transformers",
  "datasets",
  "scikit-learn",
  "spacy",
  "imbalanced-learn",
  "swifter",
  "huggingface_hub",
  "pytorch-lightning",
  "matplotlib",
  "scipy",
  "joblib",
  "pydeps",
  "isort",
  "URLExtract",
  "emoji>=2.0.0",
  "autocorrect",
  "spello",
  "pycaleva @ https://github.com/itchingpixels/pycaleva/archive/patch-1.zip",
  "perspective",
]

[tool.hatch.metadata]
allow-direct-references = true

[project.urls]
"Homepage" = "https://github.com/applied-exploration/modular-pipelines"
"Bug Tracker" = "https://github.com/applied-exploration/modular-pipelines/issues"